feat(editor): document-centric selection/delete/export

+++ editor/qt/include/panels/property_panel.hpp
@@ -3,6 +3,7 @@
 #include <QDockWidget>
 #include <QList>
 #include <QVariant>
+#include <QtGlobal>
 class QTreeWidget;
 class QCheckBox;
 
@@ -11,16 +12,16 @@ class PropertyPanel : public QDockWidget {
 public:
     explicit PropertyPanel(QWidget* parent = nullptr);
     ~PropertyPanel() override;
-    void updateFromSelection(const QList<int>& entityIds);
+    void updateFromSelection(const QList<qulonglong>& entityIds);
     void setVisibleCheckState(Qt::CheckState state, bool silent=true);
 
 signals:
-    void propertyEdited(int entityId, const QString& key, const QVariant& value);
-    void propertyEditedBatch(const QList<int>& entityIds, const QString& key, const QVariant& value);
+    void propertyEdited(qulonglong entityId, const QString& key, const QVariant& value);
+    void propertyEditedBatch(const QList<qulonglong>& entityIds, const QString& key, const QVariant& value);
 
 private:
     QTreeWidget* m_tree{nullptr};
-    QList<int> m_currentSelection;
+    QList<qulonglong> m_currentSelection;
     QCheckBox* m_visibleCheck{nullptr};
     bool m_internalChange{false};
 };
